requirements.txt
aws-cdk-lib==2.92.0
aws-cdk.aws-batch-alpha==2.92.0a0
constructs>=10.0.0,<11.0.0
